Fully implement improved Fractal Voronoi noise

+++ source/blender/gpu/shaders/material/gpu_shader_material_fractal_voronoi.glsl
@@ -26,7 +26,7 @@ void fractal_voronoi_f1(float coord,
 
   max_amplitude = 0.0;
   outDistance = 0.0;
-  for (int i = 0; i <= ceilf(detail); ++i) {
+  for (int i = 0; i <= ceil(detail); ++i) {
     voronoi_f1(coord * octave_scale, randomness, octave_distance, octave_color, octave_postion);
     if (detail == 0.0 || roughness == 0.0 || lacunarity == 0.0) {
       max_amplitude = 1.0;
@@ -44,7 +44,7 @@ void fractal_voronoi_f1(float coord,
       octave_amplitude *= roughness;
     }
     else {
-      float remainder = detail - floorf(detail);
+      float remainder = detail - floor(detail);
       if (remainder != 0.0) {
         max_amplitude = mix(max_amplitude, max_amplitude + octave_amplitude, remainder);
         outDistance = mix(
@@ -77,7 +77,7 @@ void fractal_voronoi_smooth_f1(float coord,
 
   max_amplitude = 0.0;
   outDistance = 0.0;
-  for (int i = 0; i <= ceilf(detail); ++i) {
+  for (int i = 0; i <= ceil(detail); ++i) {
     voronoi_smooth_f1(coord * octave_scale,
                       smoothness,
                       randomness,
@@ -100,7 +100,7 @@ void fractal_voronoi_smooth_f1(float coord,
       octave_amplitude *= roughness;
     }
     else {
-      float remainder = detail - floorf(detail);
+      float remainder = detail - floor(detail);
       if (remainder != 0.0) {
         max_amplitude = mix(max_amplitude, max_amplitude + octave_amplitude, remainder);
         outDistance = mix(
@@ -132,7 +132,7 @@ void fractal_voronoi_f2(float coord,
 
   max_amplitude = 0.0;
   outDistance = 0.0;
-  for (int i = 0; i <= ceilf(detail); ++i) {
+  for (int i = 0; i <= ceil(detail); ++i) {
     voronoi_f2(coord * octave_scale, randomness, octave_distance, octave_color, octave_postion);
     if (detail == 0.0 || roughness == 0.0 || lacunarity == 0.0) {
       max_amplitude = 1.0;
@@ -150,7 +150,7 @@ void fractal_voronoi_f2(float coord,
       octave_amplitude *= roughness;
     }
     else {
-      float remainder = detail - floorf(detail);
+      float remainder = detail - floor(detail);
       if (remainder != 0.0) {
         max_amplitude = mix(max_amplitude, max_amplitude + octave_amplitude, remainder);
         outDistance = mix(
@@ -172,30 +172,30 @@ void fractal_voronoi_distance_to_edge(float coord,
                                       out float max_amplitude,
                                       out float outDistance)
 {
-  float octave_scale = 1.0f;
-  float octave_amplitude = 1.0f;
-  float octave_distance = 0.0f;
+  float octave_scale = 1.0;
+  float octave_amplitude = 1.0;
+  float octave_distance = 0.0;
 
-  max_amplitude = 2.0f - randomness;
-  outDistance = 8.0f;
-  for (int i = 0; i <= ceilf(detail); ++i) {
+  max_amplitude = 2.0 - randomness;
+  outDistance = 8.0;
+  for (int i = 0; i <= ceil(detail); ++i) {
     voronoi_distance_to_edge(coord * octave_scale, randomness, octave_distance);
-    if (detail == 0.0f || roughness == 0.0f || lacunarity == 0.0f) {
+    if (detail == 0.0 || roughness == 0.0 || lacunarity == 0.0) {
       outDistance = octave_distance;
       return;
     }
     else if (i <= detail) {
-      max_amplitude = mix(max_amplitude, (2.0f - randomness) * octave_scale, octave_amplitude);
+      max_amplitude = mix(max_amplitude, (2.0 - randomness) * octave_scale, octave_amplitude);
       outDistance = mix(
           outDistance, min(outDistance, octave_distance / octave_scale), octave_amplitude);
       octave_scale *= lacunarity;
       octave_amplitude *= roughness;
     }
     else {
-      float remainder = detail - floorf(detail);
-      if (remainder != 0.0f) {
+      float remainder = detail - floor(detail);
+      if (remainder != 0.0) {
         float lerp_amplitude = mix(
-            max_amplitude, (2.0f - randomness) * octave_scale, octave_amplitude);
+            max_amplitude, (2.0 - randomness) * octave_scale, octave_amplitude);
         max_amplitude = mix(max_amplitude, lerp_amplitude, remainder);
         float lerp_distance = mix(
             outDistance, min(outDistance, octave_distance / octave_scale), octave_amplitude);
@@ -227,7 +227,7 @@ void fractal_voronoi_f1(vec2 coord,
 
   max_amplitude = 0.0;
   outDistance = 0.0;
-  for (int i = 0; i <= ceilf(detail); ++i) {
+  for (int i = 0; i <= ceil(detail); ++i) {
     voronoi_f1(coord * octave_scale,
                exponent,
                randomness,
@@ -251,7 +251,7 @@ void fractal_voronoi_f1(vec2 coord,
       octave_amplitude *= roughness;
     }
     else {
-      float remainder = detail - floorf(detail);
+      float remainder = detail - floor(detail);
       if (remainder != 0.0) {
         max_amplitude = mix(max_amplitude, max_amplitude + octave_amplitude, remainder);
         outDistance = mix(
@@ -286,7 +286,7 @@ void fractal_voronoi_smooth_f1(vec2 coord,
 
   max_amplitude = 0.0;
   outDistance = 0.0;
-  for (int i = 0; i <= ceilf(detail); ++i) {
+  for (int i = 0; i <= ceil(detail); ++i) {
     voronoi_smooth_f1(coord * octave_scale,
                       smoothness,
                       exponent,
@@ -311,7 +311,7 @@ void fractal_voronoi_smooth_f1(vec2 coord,
       octave_amplitude *= roughness;
     }
     else {
-      float remainder = detail - floorf(detail);
+      float remainder = detail - floor(detail);
       if (remainder != 0.0) {
         max_amplitude = mix(max_amplitude, max_amplitude + octave_amplitude, remainder);
         outDistance = mix(
@@ -345,7 +345,7 @@ void fractal_voronoi_f2(vec2 coord,
 
   max_amplitude = 0.0;
   outDistance = 0.0;
-  for (int i = 0; i <= ceilf(detail); ++i) {
+  for (int i = 0; i <= ceil(detail); ++i) {
     voronoi_f2(coord * octave_scale,
                exponent,
                randomness,
@@ -369,7 +369,7 @@ void fractal_voronoi_f2(vec2 coord,
       octave_amplitude *= roughness;
     }
     else {
-      float remainder = detail - floorf(detail);
+      float remainder = detail - floor(detail);
       if (remainder != 0.0) {
         max_amplitude = mix(max_amplitude, max_amplitude + octave_amplitude, remainder);
         outDistance = mix(
@@ -391,30 +391,30 @@ void fractal_voronoi_distance_to_edge(vec2 coord,
                                       out float max_amplitude,
                                       out float outDistance)
 {
-  float octave_scale = 1.0f;
-  float octave_amplitude = 1.0f;
-  float octave_distance = 0.0f;
+  float octave_scale = 1.0;
+  float octave_amplitude = 1.0;
+  float octave_distance = 0.0;
 
-  max_amplitude = 2.0f - randomness;
-  outDistance = 8.0f;
-  for (int i = 0; i <= ceilf(detail); ++i) {
+  max_amplitude = 2.0 - randomness;
+  outDistance = 8.0;
+  for (int i = 0; i <= ceil(detail); ++i) {
     voronoi_distance_to_edge(coord * octave_scale, randomness, octave_distance);
-    if (detail == 0.0f || roughness == 0.0f || lacunarity == 0.0f) {
+    if (detail == 0.0 || roughness == 0.0 || lacunarity == 0.0) {
       outDistance = octave_distance;
       return;
     }
     else if (i <= detail) {
-      max_amplitude = mix(max_amplitude, (2.0f - randomness) * octave_scale, octave_amplitude);
+      max_amplitude = mix(max_amplitude, (2.0 - randomness) * octave_scale, octave_amplitude);
       outDistance = mix(
           outDistance, min(outDistance, octave_distance / octave_scale), octave_amplitude);
       octave_scale *= lacunarity;
       octave_amplitude *= roughness;
     }
     else {
-      float remainder = detail - floorf(detail);
-      if (remainder != 0.0f) {
+      float remainder = detail - floor(detail);
+      if (remainder != 0.0) {
         float lerp_amplitude = mix(
-            max_amplitude, (2.0f - randomness) * octave_scale, octave_amplitude);
+            max_amplitude, (2.0 - randomness) * octave_scale, octave_amplitude);
         max_amplitude = mix(max_amplitude, lerp_amplitude, remainder);
         float lerp_distance = mix(
             outDistance, min(outDistance, octave_distance / octave_scale), octave_amplitude);
@@ -446,7 +446,7 @@ void fractal_voronoi_f1(vec3 coord,
 
   max_amplitude = 0.0;
   outDistance = 0.0;
-  for (int i = 0; i <= ceilf(detail); ++i) {
+  for (int i = 0; i <= ceil(detail); ++i) {
     voronoi_f1(coord * octave_scale,
                exponent,
                randomness,
@@ -470,7 +470,7 @@ void fractal_voronoi_f1(vec3 coord,
       octave_amplitude *= roughness;
     }
     else {
-      float remainder = detail - floorf(detail);
+      float remainder = detail - floor(detail);
       if (remainder != 0.0) {
         max_amplitude = mix(max_amplitude, max_amplitude + octave_amplitude, remainder);
         outDistance = mix(
@@ -505,7 +505,7 @@ void fractal_voronoi_smooth_f1(vec3 coord,
 
   max_amplitude = 0.0;
   outDistance = 0.0;
-  for (int i = 0; i <= ceilf(detail); ++i) {
+  for (int i = 0; i <= ceil(detail); ++i) {
     voronoi_smooth_f1(coord * octave_scale,
                       smoothness,
                       exponent,
@@ -530,7 +530,7 @@ void fractal_voronoi_smooth_f1(vec3 coord,
       octave_amplitude *= roughness;
     }
     else {
-      float remainder = detail - floorf(detail);
+      float remainder = detail - floor(detail);
       if (remainder != 0.0) {
         max_amplitude = mix(max_amplitude, max_amplitude + octave_amplitude, remainder);
         outDistance = mix(
@@ -564,7 +564,7 @@ void fractal_voronoi_f2(vec3 coord,
 
   max_amplitude = 0.0;
   outDistance = 0.0;
-  for (int i = 0; i <= ceilf(detail); ++i) {
+  for (int i = 0; i <= ceil(detail); ++i) {
     voronoi_f2(coord * octave_scale,
                exponent,
                randomness,
@@ -588,7 +588,7 @@ void fractal_voronoi_f2(vec3 coord,
       octave_amplitude *= roughness;
     }
     else {
-      float remainder = detail - floorf(detail);
+      float remainder = detail - floor(detail);
       if (remainder != 0.0) {
         max_amplitude = mix(max_amplitude, max_amplitude + octave_amplitude, remainder);
         outDistance = mix(
@@ -610,30 +610,30 @@ void fractal_voronoi_distance_to_edge(vec3 coord,
                                       out float max_amplitude,
                                       out float outDistance)
 {
-  float octave_scale = 1.0f;
-  float octave_amplitude = 1.0f;
-  float octave_distance = 0.0f;
+  float octave_scale = 1.0;
+  float octave_amplitude = 1.0;
+  float octave_distance = 0.0;
 
-  max_amplitude = 2.0f - randomness;
-  outDistance = 8.0f;
-  for (int i = 0; i <= ceilf(detail); ++i) {
+  max_amplitude = 2.0 - randomness;
+  outDistance = 8.0;
+  for (int i = 0; i <= ceil(detail); ++i) {
     voronoi_distance_to_edge(coord * octave_scale, randomness, octave_distance);
-    if (detail == 0.0f || roughness == 0.0f || lacunarity == 0.0f) {
+    if (detail == 0.0 || roughness == 0.0 || lacunarity == 0.0) {
       outDistance = octave_distance;
       return;
     }
     else if (i <= detail) {
-      max_amplitude = mix(max_amplitude, (2.0f - randomness) * octave_scale, octave_amplitude);
+      max_amplitude = mix(max_amplitude, (2.0 - randomness) * octave_scale, octave_amplitude);
       outDistance = mix(
           outDistance, min(outDistance, octave_distance / octave_scale), octave_amplitude);
       octave_scale *= lacunarity;
       octave_amplitude *= roughness;
     }
     else {
-      float remainder = detail - floorf(detail);
-      if (remainder != 0.0f) {
+      float remainder = detail - floor(detail);
+      if (remainder != 0.0) {
         float lerp_amplitude = mix(
-            max_amplitude, (2.0f - randomness) * octave_scale, octave_amplitude);
+            max_amplitude, (2.0 - randomness) * octave_scale, octave_amplitude);
         max_amplitude = mix(max_amplitude, lerp_amplitude, remainder);
         float lerp_distance = mix(
             outDistance, min(outDistance, octave_distance / octave_scale), octave_amplitude);
@@ -665,7 +665,7 @@ void fractal_voronoi_f1(vec4 coord,
 
   max_amplitude = 0.0;
   outDistance = 0.0;
-  for (int i = 0; i <= ceilf(detail); ++i) {
+  for (int i = 0; i <= ceil(detail); ++i) {
     voronoi_f1(coord * octave_scale,
                exponent,
                randomness,
@@ -689,7 +689,7 @@ void fractal_voronoi_f1(vec4 coord,
       octave_amplitude *= roughness;
     }
     else {
-      float remainder = detail - floorf(detail);
+      float remainder = detail - floor(detail);
       if (remainder != 0.0) {
         max_amplitude = mix(max_amplitude, max_amplitude + octave_amplitude, remainder);
         outDistance = mix(
@@ -724,7 +724,7 @@ void fractal_voronoi_smooth_f1(vec4 coord,
 
   max_amplitude = 0.0;
   outDistance = 0.0;
-  for (int i = 0; i <= ceilf(detail); ++i) {
+  for (int i = 0; i <= ceil(detail); ++i) {
     voronoi_smooth_f1(coord * octave_scale,
                       smoothness,
                       exponent,
@@ -749,7 +749,7 @@ void fractal_voronoi_smooth_f1(vec4 coord,
       octave_amplitude *= roughness;
     }
     else {
-      float remainder = detail - floorf(detail);
+      float remainder = detail - floor(detail);
       if (remainder != 0.0) {
         max_amplitude = mix(max_amplitude, max_amplitude + octave_amplitude, remainder);
         outDistance = mix(
@@ -783,7 +783,7 @@ void fractal_voronoi_f2(vec4 coord,
 
   max_amplitude = 0.0;
   outDistance = 0.0;
-  for (int i = 0; i <= ceilf(detail); ++i) {
+  for (int i = 0; i <= ceil(detail); ++i) {
     voronoi_f2(coord * octave_scale,
                exponent,
                randomness,
@@ -807,7 +807,7 @@ void fractal_voronoi_f2(vec4 coord,
       octave_amplitude *= roughness;
     }
     else {
-      float remainder = detail - floorf(detail);
+      float remainder = detail - floor(detail);
       if (remainder != 0.0) {
         max_amplitude = mix(max_amplitude, max_amplitude + octave_amplitude, remainder);
         outDistance = mix(
@@ -829,30 +829,30 @@ void fractal_voronoi_distance_to_edge(vec4 coord,
                                       out float max_amplitude,
                                       out float outDistance)
 {
-  float octave_scale = 1.0f;
-  float octave_amplitude = 1.0f;
-  float octave_distance = 0.0f;
+  float octave_scale = 1.0;
+  float octave_amplitude = 1.0;
+  float octave_distance = 0.0;
 
-  max_amplitude = 2.0f - randomness;
-  outDistance = 8.0f;
-  for (int i = 0; i <= ceilf(detail); ++i) {
+  max_amplitude = 2.0 - randomness;
+  outDistance = 8.0;
+  for (int i = 0; i <= ceil(detail); ++i) {
     voronoi_distance_to_edge(coord * octave_scale, randomness, octave_distance);
-    if (detail == 0.0f || roughness == 0.0f || lacunarity == 0.0f) {
+    if (detail == 0.0 || roughness == 0.0 || lacunarity == 0.0) {
       outDistance = octave_distance;
       return;
     }
     else if (i <= detail) {
-      max_amplitude = mix(max_amplitude, (2.0f - randomness) * octave_scale, octave_amplitude);
+      max_amplitude = mix(max_amplitude, (2.0 - randomness) * octave_scale, octave_amplitude);
       outDistance = mix(
           outDistance, min(outDistance, octave_distance / octave_scale), octave_amplitude);
       octave_scale *= lacunarity;
       octave_amplitude *= roughness;
     }
     else {
-      float remainder = detail - floorf(detail);
-      if (remainder != 0.0f) {
+      float remainder = detail - floor(detail);
+      if (remainder != 0.0) {
         float lerp_amplitude = mix(
-            max_amplitude, (2.0f - randomness) * octave_scale, octave_amplitude);
+            max_amplitude, (2.0 - randomness) * octave_scale, octave_amplitude);
         max_amplitude = mix(max_amplitude, lerp_amplitude, remainder);
         float lerp_distance = mix(
             outDistance, min(outDistance, octave_distance / octave_scale), octave_amplitude);
